Former-commit-id: ed30fb412d86f9727d72d4fb532ab144d2ea234e

+++ libraries/reconstruction/image_sort_by_statistics.cpp
@@ -90,7 +90,7 @@ void ProgSortByStatistics::processInprocessInputPrepareSPTH(MetaData &SF)
     //Histogram analysis, to detect black points and saturated parts
     tempPcaAnalyzer4.clear();
 
-    int sign = 1;//;-1;
+    double sign = 1;//;-1;
     int numNorm = 3;
     int numDescriptors0=numNorm;
     int numDescriptors2=4;
